Move `NativeLibKind`
