+++ compiler/rustc_attr_ir/src/data_structures.rs
@@ -16,7 +16,7 @@ use rustc_macros::{Decodable, Encodable, PrintAttribute, StableHash};
 use rustc_span::def_id::DefId;
 use rustc_span::hygiene::Transparency;
 use rustc_span::{ErrorGuaranteed, Ident, Span, Symbol};
-use rustc_structures::{CrateType, Limit, SanitizerSet};
+use rustc_structures::{CrateType, Limit, NativeLibKind, SanitizerSet};
 use thin_vec::ThinVec;
 
 pub use crate::canonical_symbols::{CanonicalSymbol, CanonicalSymbols};
@@ -369,76 +369,6 @@ pub enum PeImportNameType {
     /// Prefix (e.g., the leading `_` or `@`) and suffix (the first `@` and all
     /// trailing characters) are skipped.
     Undecorated,
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-#[derive(Encodable, Decodable, PrintAttribute)]
-#[derive(StableHash)]
-pub enum NativeLibKind {
-    /// Static library (e.g. `libfoo.a` on Linux or `foo.lib` on Windows/MSVC)
-    Static {
-        /// Whether to bundle objects from static library into produced rlib
-        bundle: Option<bool>,
-        /// Whether to link static library without throwing any object files away
-        whole_archive: Option<bool>,
-        /// Whether to export c static library symbols
-        export_symbols: Option<bool>,
-    },
-    /// Dynamic library (e.g. `libfoo.so` on Linux)
-    /// or an import library corresponding to a dynamic library (e.g. `foo.lib` on Windows/MSVC).
-    Dylib {
-        /// Whether the dynamic library will be linked only if it satisfies some undefined symbols
-        as_needed: Option<bool>,
-    },
-    /// Dynamic library (e.g. `foo.dll` on Windows) without a corresponding import library.
-    /// On Linux, it refers to a generated shared library stub.
-    RawDylib {
-        /// Whether the dynamic library will be linked only if it satisfies some undefined symbols
-        as_needed: Option<bool>,
-    },
-    /// A macOS-specific kind of dynamic libraries.
-    Framework {
-        /// Whether the framework will be linked only if it satisfies some undefined symbols
-        as_needed: Option<bool>,
-    },
-    /// Argument which is passed to linker, relative order with libraries and other arguments
-    /// is preserved
-    LinkArg,
-
-    /// Module imported from WebAssembly
-    WasmImportModule,
-
-    /// The library kind wasn't specified, `Dylib` is currently used as a default.
-    Unspecified,
-}
-
-impl NativeLibKind {
-    pub fn has_modifiers(&self) -> bool {
-        match self {
-            NativeLibKind::Static { bundle, whole_archive, export_symbols } => {
-                bundle.is_some() || whole_archive.is_some() || export_symbols.is_some()
-            }
-            NativeLibKind::Dylib { as_needed }
-            | NativeLibKind::Framework { as_needed }
-            | NativeLibKind::RawDylib { as_needed } => as_needed.is_some(),
-            NativeLibKind::Unspecified
-            | NativeLibKind::LinkArg
-            | NativeLibKind::WasmImportModule => false,
-        }
-    }
-
-    pub fn is_statically_included(&self) -> bool {
-        matches!(self, NativeLibKind::Static { .. })
-    }
-
-    pub fn is_dllimport(&self) -> bool {
-        matches!(
-            self,
-            NativeLibKind::Dylib { .. }
-                | NativeLibKind::RawDylib { .. }
-                | NativeLibKind::Unspecified
-        )
-    }
 }
 
 #[derive(Debug, Encodable, Decodable, Clone, StableHash, PrintAttribute)]
