+++ compiler/rustc_codegen_ssa/src/back/link/raw_dylib.rs
@@ -8,9 +8,9 @@ use rustc_data_structures::base_n::{CASE_INSENSITIVE, ToBaseN};
 use rustc_data_structures::fx::{FxHashMap, FxIndexMap};
 use rustc_data_structures::stable_hash::StableHasher;
 use rustc_hashes::Hash128;
-use rustc_hir::attrs::NativeLibKind;
 use rustc_session::Session;
 use rustc_span::Symbol;
+use rustc_structures::NativeLibKind;
 use rustc_target::spec::Arch;
 
 use crate::back::archive::ImportLibraryItem;
