+++ compiler/rustc_crate_store/Cargo.toml
@@ -12,4 +12,5 @@ rustc_hir_id = { path = "../rustc_hir_id" }
 rustc_macros = { path = "../rustc_macros" }
 rustc_serialize = { path = "../rustc_serialize" }
 rustc_span = { path = "../rustc_span" }
+rustc_structures = { path = "../rustc_structures" }
 # tidy-alphabetical-end
