+++ compiler/rustc_crate_store/src/cstore.rs
@@ -6,7 +6,7 @@ use std::any::Any;
 use std::path::PathBuf;
 
 use rustc_abi::ExternAbi;
-use rustc_attr_ir::{CfgEntry, NativeLibKind, PeImportNameType};
+use rustc_attr_ir::{CfgEntry, PeImportNameType};
 use rustc_data_structures::sync::{self, AppendOnlyIndexVec, FreezeLock};
 use rustc_hir_id::definitions::{DefKey, DefPath, Definitions};
 use rustc_macros::{BlobDecodable, Decodable, Encodable, StableHash};
@@ -14,6 +14,7 @@ use rustc_span::def_id::{
     CrateNum, DefId, DefPathHash, LOCAL_CRATE, LocalDefId, StableCrateId, StableCrateIdMap,
 };
 use rustc_span::{Span, Symbol};
+use rustc_structures::NativeLibKind;
 
 /// Where a crate came from on the local filesystem. One of these three options
 /// must be non-None.
