+++ compiler/rustc_metadata/src/native_libs.rs
@@ -4,7 +4,7 @@ use rustc_crate_store::{
     DllCallingConvention, DllImport, DllImportSymbolType, ForeignModule, NativeLib,
 };
 use rustc_data_structures::fx::FxHashSet;
-use rustc_hir::attrs::{NativeLibKind, PeImportNameType};
+use rustc_hir::attrs::PeImportNameType;
 use rustc_hir::def::DefKind;
 use rustc_hir::find_attr;
 use rustc_middle::bug;
@@ -14,6 +14,7 @@ use rustc_middle::ty::{self, List, Ty, TyCtxt};
 use rustc_session::Session;
 use rustc_span::Symbol;
 use rustc_span::def_id::{DefId, LOCAL_CRATE};
+use rustc_structures::NativeLibKind;
 use rustc_target::spec::{Arch, BinaryFormat, CfgAbi};
 
 use crate::diagnostics;
