+++ compiler/rustc_session/src/utils.rs
@@ -1,10 +1,10 @@
 use std::path::PathBuf;
 use std::sync::OnceLock;
 
-use rustc_attr_ir::NativeLibKind;
 use rustc_data_structures::profiling::VerboseTimingGuard;
 use rustc_fs_util::try_canonicalize;
 use rustc_macros::{Decodable, Encodable, StableHash};
+use rustc_structures::NativeLibKind;
 
 use crate::session::Session;
