+++ compiler/rustc_structures/src/lib.rs
@@ -2,8 +2,10 @@
 
 mod crate_type;
 mod limit;
+mod native_lib_kind;
 mod sanitizer_set;
 
 pub use crate_type::CrateType;
 pub use limit::Limit;
+pub use native_lib_kind::NativeLibKind;
 pub use sanitizer_set::SanitizerSet;
